+++ client/src/pages/AuthorizedLayer/index.tsx
@@ -1,6 +1,5 @@
 import FullScreenSpinner from '@components/FullScreenSpinner';
-import { useMyInfoQueryData } from '@hooks/auth';
-import useReissueTokenMutation from '@hooks/useReissueTokenMutation';
+import { useMyInfoQueryData, useReissueTokenMutation } from '@hooks/auth';
 import { useTokenStore } from '@stores/tokenStore';
 import React, { useEffect } from 'react';
 import { Outlet, useNavigate } from 'react-router-dom';
